+++ toggl_api/__init__.py
@@ -2,26 +2,21 @@ from .client import ClientBody, ClientEndpoint
 from .config import generate_authentication
 from .meta import JSONCache
 from .meta.cache import Comparison, TogglQuery
-from .models import (
-    TogglClient,
-    TogglProject,
-    TogglTag,
-    TogglTracker,
-    TogglWorkspace,
-)
+from .models import TogglClient, TogglOrganization, TogglProject, TogglTag, TogglTracker, TogglWorkspace
+from .organization import OrganizationEndpoint
 from .project import ProjectBody, ProjectEndpoint
 from .tag import TagEndpoint
 from .tracker import TrackerBody, TrackerEndpoint
 from .user import UserEndpoint
 from .version import version
-from .workspace import WorkspaceEndpoint
+from .workspace import WorkspaceBody, WorkspaceEndpoint
 
 __author__ = "David Kasakaitis"
 __version__ = version
 __typed__ = True
 
 
-__all__ = [
+__all__ = (  # noqa: RUF022
     "ClientBody",
     "ClientEndpoint",
     "Comparison",
@@ -39,6 +34,9 @@ __all__ = [
     "TrackerEndpoint",
     "UserEndpoint",
     "UserEndpoint",
+    "WorkspaceBody",
     "WorkspaceEndpoint",
     "generate_authentication",
-]
+    "TogglOrganization",
+    "OrganizationEndpoint",
+)